+++ Libft/ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: tsongtra <tsongtra@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/25 18:41:02 by tsongtra          #+#    #+#             */
-/*   Updated: 2023/08/26 00:00:10 by tsongtra         ###   ########.fr       */
+/*   Updated: 2023/08/27 02:24:31 by tsongtra         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n)
